backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
pytesseract
pillow
google-generativeai